+++ src/constraint_system/composer.rs
@@ -209,6 +209,11 @@ impl StandardComposer {
         composer
     }
 
+    /// Witness representation of zero of the first variable of any circuit
+    pub const fn zero_var(&self) -> Variable {
+        self.zero_var
+    }
+
     /// Add Input first calls the Permutation
     /// to generate and allocate a new [`Variable`] `var`.
     ///
@@ -655,7 +660,7 @@ mod tests {
         let res = gadget_tester(
             |composer| {
                 let bit_1 = composer.add_input(BlsScalar::one());
-                let bit_0 = composer.add_input(BlsScalar::zero());
+                let bit_0 = composer.zero_var();
 
                 let choice_a = composer.add_input(BlsScalar::from(10u64));
                 let choice_b = composer.add_input(BlsScalar::from(20u64));